Create separate config package
* Delete imports package
