+++ BUILD
@@ -6,13 +6,13 @@ python_binary(
         "//adapters/plz_cli",
         "//common",
         "//common/logger",
+        "//config",
         "//domain/build_pkgs",
-        "//domain/imports",
-        "//domain/imports/known",
-        "//domain/imports/stdlib",
         "//domain/plz/target",
+        "//domain/python_import/stdlib",
         "//service/ast/converters",
         "//service/dependency",
+        "//service/python_import:imports",
         "//third_party/python3:colorama",
     ],
 )
@@ -25,13 +25,13 @@ python_binary(
         "//adapters/plz_cli",
         "//common",
         "//common/logger",
+        "//config",
         "//domain/build_pkgs",
-        "//domain/imports",
-        "//domain/imports/known",
-        "//domain/imports/stdlib",
         "//domain/plz/target",
+        "//domain/python_import/stdlib",
         "//service/ast/converters",
         "//service/dependency",
+        "//service/python_import:imports",
         "//third_party/python3:colorama",
     ],
 )
