+++ config/BUILD
@@ -1,7 +1,7 @@
-package(default_visibility=["PUBLIC"])
+package(default_visibility = ["PUBLIC"])
 
 python_library(
-    name = "known",
+    name = "config",
     srcs = glob(
         ["*.py"],
         exclude = ["*_test.py"],
@@ -14,11 +14,11 @@ python_library(
 )
 
 python_test(
-    name = "known_test",
+    name = "config_test",
     srcs = glob(["*_test.py"]),
     deps = [
-        ":known",
-        "//domain/targets",
+        ":config",
+        "//domain/plz/target",
         "//third_party/python3:jsonschema",
     ],
 )
