+++ config/common.py
@@ -1,0 +1,3 @@
+from common.logger.logger import setup_logger
+
+LOGGER = setup_logger(__file__)
