--- domain/imports/BUILD
@@ -1,9 +1,0 @@
-package(default_visibility = ["PUBLIC"])
-
-python_library(
-    name = "imports",
-    srcs = glob(
-        ["*.py"],
-        exclude = ["*_test.py"],
-    ),
-)
