+++ domain/python_import/stdlib/2_7.py
